fix: reverting user doctype override
