+++ lms/lms/utils.py
@@ -1,6 +1,7 @@
 import re
 import string
 import frappe
+import hashlib
 import json
 import razorpay
 import requests
@@ -2009,3 +2010,25 @@ def get_batch_card_details(batches):
 			batch.price = fmt_money(batch.amount, 0, batch.currency)
 
 	return batches
+
+
+def get_palette(full_name):
+	"""
+	Returns a color unique to each member for Avatar"""
+
+	palette = [
+		["--orange-avatar-bg", "--orange-avatar-color"],
+		["--pink-avatar-bg", "--pink-avatar-color"],
+		["--blue-avatar-bg", "--blue-avatar-color"],
+		["--green-avatar-bg", "--green-avatar-color"],
+		["--dark-green-avatar-bg", "--dark-green-avatar-color"],
+		["--red-avatar-bg", "--red-avatar-color"],
+		["--yellow-avatar-bg", "--yellow-avatar-color"],
+		["--purple-avatar-bg", "--purple-avatar-color"],
+		["--gray-avatar-bg", "--gray-avatar-color0"],
+	]
+
+	encoded_name = str(full_name).encode("utf-8")
+	hash_name = hashlib.md5(encoded_name).hexdigest()
+	idx = cint((int(hash_name[4:6], 16) + 1) / 5.33)
+	return palette[idx % 8]
