--- lms/lms/web_template/courses_mentored/courses_mentored.json
@@ -1,16 +1,0 @@
-{
- "__unsaved": 1,
- "creation": "2021-10-21 11:32:57.411626",
- "docstatus": 0,
- "doctype": "Web Template",
- "fields": [],
- "idx": 0,
- "modified": "2021-10-21 12:01:56.270656",
- "modified_by": "Administrator",
- "module": "LMS",
- "name": "Courses Mentored",
- "owner": "Administrator",
- "standard": 1,
- "template": "",
- "type": "Section"
-}
