+++ lms/templates/signup-form.html
@@ -85,7 +85,7 @@
         }
 
         frappe.call({
-            method: "lms.overrides.user.sign_up",
+            method: "lms.lms.user.sign_up",
             args: {
                 "email": email,
                 "full_name": full_name,
